+++ test/integration/offermaker.integration.test.ts
@@ -51,7 +51,7 @@ describe("OfferMaker integration test suite", () => {
       base: "TokenA",
       quote: "TokenB",
       tickSpacing: 1,
-      bookOptions: { maxOffers: 30 },
+      bookOptions: { targetNumberOfTicks: 30 },
     });
     onchain_lp = await LiquidityProvider.connect(logic, 20000, market);
     eoa_lp = await mgv.liquidityProvider(market);